fix tests to match new behavior
We are now pulling the unit id from the initial login response,
instead of making users figure that out for themselves. This is less
error prone, though a little less flexible, and means multi unit
accounts only get the first one for now.

We'll solve that when there are more than 1 user on the platform.

This is also updating README to be more real